pyproject.toml
[tool.poetry]
name = "slack-chatbot-llm"
version = "0.1.0"
description = ""
authors = ["yunozon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
langchain = "^0.3.7"
openai = "^1.54.3"
slack-sdk = "^3.33.3"
slack-bolt = "^1.21.2"
flask = "^3.0.3"
langchain-openai = "^0.2.9"
langchain-google-genai = "^2.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
